+++ packages/desktop-app/src/renderer/domains/settings/components/McpSettings.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, useImperativeHandle } from "react";
 import { ManualMcpDialog } from "./ManualMcpDialog";
-import { McpEditDrawer } from "./McpEditDrawer";
+import { McpEditDialog } from "./McpEditDialog";
 import { McpSettingsView } from "./McpSettingsView";
 import { useMcpSettingsModel } from "./useMcpSettingsModel";
 
@@ -28,7 +28,7 @@ export const McpSettings = forwardRef<McpSettingsHandle>(function McpSettings(_p
 		<>
 			<McpSettingsView model={model} />
 			<ManualMcpDialog model={model} />
-			<McpEditDrawer model={model} />
+			<McpEditDialog model={model} />
 		</>
 	);
 });